requirements.txt
aiogram==3.2.0
requests==2.32.3
beautifulsoup4==4.12.3
apscheduler==3.10.4
gspread==6.1.0
google-auth==2.28.2
python-dotenv==1.0.1
